cli/requirements.txt
attrs==17.4.0
certifi==2018.1.18
certstream==1.9
chardet==3.0.4
coverage==4.5.1
idna==2.6
mock==2.0.0
mongomock==3.9.0
more-itertools==4.1.0
numpy==1.14.2
pandas==0.22.0
pbr==4.0.2
pluggy==0.6.0
py==1.5.3
pymongo==3.6.1
pytest==3.5.0
pytest-cov==2.5.1
python-dateutil==2.7.0
python-Levenshtein==0.12.0
pytz==2018.3
PyYAML==3.12
requests==2.18.4
requests-file==1.4.3
scikit-learn==0.19.1
scipy==1.0.0
sentinels==1.0.0
six==1.11.0
sklearn==0.0
termcolor==1.1.0
tldextract==2.2.0
tqdm==4.19.8
urllib3==1.22
websocket-client==0.47.0
